+++ common/Parser/MapParser/MapParser.cpp
@@ -165,8 +165,8 @@ void MapParser::parseWaves(const nlohmann::json& waves, float tileWidth) {
 
     std::random_device rd;
     std::mt19937 gen(rd());
-    std::uniform_real_distribution<> disY(0.0f, static_cast<float>(constants::MAX_HEIGHT));
-    std::uniform_real_distribution<> disX(0.0f, 1.0f);
+    std::uniform_real_distribution<> disY(0.0, static_cast<double>(constants::MAX_HEIGHT));
+    std::uniform_real_distribution<> disX(0.0, 1.0);
 
     for (size_t waveIndex = 0; waveIndex < waves.size(); ++waveIndex) {
         const auto& wave = waves[waveIndex];
